refactor(functional): move pipe function, rename old pipe to flow
