+++ src/utils/index.ts
@@ -326,40 +326,3 @@ export function normalizeForStore<T, K extends keyof T>(objects: T[], key: K) {
 
   return { order, byKey };
 }
-
-// Pipe code from
-// https://dev.to/ecyrbe/how-to-use-advanced-typescript-to-define-a-pipe-function-381h
-type AnyFunc = (...arg: any) => any;
-
-type LastFnReturnType<F extends Array<AnyFunc>, Else = never> = F extends [
-  ...any[],
-  (...arg: any) => infer R
-]
-  ? R
-  : Else;
-
-type PipeArgs<F extends AnyFunc[], Acc extends AnyFunc[] = []> = F extends [
-  (...args: infer A) => infer B
-]
-  ? [...Acc, (...args: A) => B]
-  : F extends [(...args: infer A) => any, ...infer Tail]
-  ? Tail extends [(arg: infer B) => any, ...any[]]
-    ? PipeArgs<Tail, [...Acc, (...args: A) => B]>
-    : Acc
-  : Acc;
-
-// Example:
-// const valid = pipe(
-//   "1",
-//   (a: string) => Number(a),
-//   (c: number) => c + 1,
-//   (d: number) => `${d}`,
-//   (e: string) => Number(e)
-// );
-export function pipe<FirstFn extends AnyFunc, F extends AnyFunc[]>(
-  arg: Parameters<FirstFn>[0],
-  firstFn: FirstFn,
-  ...fns: PipeArgs<F> extends F ? F : PipeArgs<F>
-): LastFnReturnType<F, ReturnType<FirstFn>> {
-  return (fns as AnyFunc[]).reduce((acc, fn) => fn(acc), firstFn(arg));
-}
